requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
djangorestframework-simplejwt==4.7.2
